+++ assets/src/components/editing/commands/AudioCmd.tsx
@@ -8,6 +8,7 @@ import { MediaManager } from 'components/media/manager/MediaManager.controller';
 import { modalActions } from 'actions/modal';
 import { MediaItem } from 'types/media';
 import { Command, CommandDesc } from 'components/editing/commands/interfaces';
+import guid from 'utils/guid';
 
 const dismiss = () => (window as any).oliDispatch(modalActions.dismiss());
 const display = (c: any) => (window as any).oliDispatch(modalActions.display(c));
@@ -32,7 +33,7 @@ export function selectAudio(
         <MediaManager
           projectSlug={projectSlug}
           // eslint-disable-next-line
-          onEdit={() => { }}
+          onEdit={() => {}}
           mimeFilter={MIMETYPE_FILTERS.AUDIO}
           selectionType={SELECTION_TYPES.SINGLE}
           initialSelectionPaths={[model.src]}
@@ -47,7 +48,7 @@ export function selectAudio(
   });
 }
 
-const command: Command = {
+const libraryCommand: Command = {
   execute: (context, editor: ReactEditor) => {
     const at = editor.selection as any;
     selectAudio(context.projectSlug, ContentModel.audio()).then((img) =>
@@ -59,9 +60,44 @@ const command: Command = {
   },
 };
 
-export const commandDesc: CommandDesc = {
-  type: 'CommandDesc',
-  icon: () => 'audiotrack',
-  description: () => 'Audio Clip',
-  command,
-};
+function createCustomEventCommand(onRequestMedia: (r: any) => Promise<string | boolean>) {
+  const customEventCommand: Command = {
+    execute: (context, editor: ReactEditor) => {
+      const at = editor.selection as any;
+
+      const request = {
+        type: 'MediaItemRequest',
+        mimeTypes: MIMETYPE_FILTERS.AUDIO,
+      };
+
+      onRequestMedia(request).then((r) => {
+        if (typeof r === 'string') {
+          const img = {
+            type: 'img',
+            src: r as string,
+            id: guid(),
+            children: [],
+          };
+          Transforms.insertNodes(editor, img, { at });
+        }
+      });
+    },
+    precondition: (editor: ReactEditor) => {
+      return true;
+    },
+  };
+  return customEventCommand;
+}
+
+export function getCommand(onRequestMedia: any) {
+  const commandDesc: CommandDesc = {
+    type: 'CommandDesc',
+    icon: () => 'audiotrack',
+    description: () => 'Audio Clip',
+    command:
+      onRequestMedia === null || onRequestMedia === undefined
+        ? libraryCommand
+        : createCustomEventCommand(onRequestMedia),
+  };
+  return commandDesc;
+}