+++ assets/src/components/resource/editors/createEditor.tsx
@@ -8,29 +8,15 @@ import {
   ResourceType,
 } from 'data/content/resource';
 import { StructuredContentEditor } from 'components/content/StructuredContentEditor';
-import { EditorDesc } from 'data/content/editors';
 import { ContentBlock } from './ContentBlock';
 import { ActivityBlock } from './ActivityBlock';
 import { getToolbarForResourceType } from '../../editing/toolbars/insertion/items';
-import { UnsupportedActivity } from '../UnsupportedActivity';
 import * as Immutable from 'immutable';
 import { defaultState } from '../TestModeHandler';
 import { ActivityEditContext } from 'data/content/activity';
 import { InlineActivityEditor, EditorUpdate } from 'components/activity/InlineActivityEditor';
 import { Objective } from 'data/content/objective';
 import { Undoable } from 'components/activities/types';
-
-const unsupported: EditorDesc = {
-  deliveryElement: UnsupportedActivity,
-  authoringElement: UnsupportedActivity,
-  icon: '',
-  description: 'Not supported',
-  friendlyName: 'Not supported',
-  slug: 'unknown',
-  globallyAvailable: true,
-  enabledForProject: true,
-  id: -1,
-};
 
 // content or referenced activities
 export const createEditor = (
@@ -61,6 +47,7 @@ export const createEditor = (
           projectSlug={projectSlug}
           toolbarItems={getToolbarForResourceType(
             graded ? ResourceType.assessment : ResourceType.page,
+            null,
           )}
         />
       </ContentBlock>