+++ lib/oli/analytics/datashop/elements/event_descriptor.ex
@@ -16,21 +16,12 @@ defmodule Oli.Analytics.Datashop.Elements.EventDescriptor do
   require Logger
   alias Oli.Analytics.Datashop.Utils
 
-  def setup(%{type: type, problem_name: problem_name, part_attempt: part_attempt}) do
+  def setup(type, %{problem_name: problem_name, part_attempt: part_attempt}) do
     element(:event_descriptor, [
       element(:selection, problem_name),
-      element(:action, get_action(part_attempt)),
+      element(:action, "Activity submission"),
       element(:input, get_input(type, part_attempt))
     ])
-  end
-
-  defp get_action(part_attempt) do
-    case activity_type(part_attempt) do
-      "oli_short_answer" -> "Short answer input"
-      "oli_multiple_choice" -> "Multiple choice selection"
-      "oli_check_all_that_apply" -> "Check all that apply selection"
-      _unregistered -> "Action in unregistered activity type"
-    end
   end
 
   defp get_input(type, part_attempt) do
@@ -43,60 +34,25 @@ defmodule Oli.Analytics.Datashop.Elements.EventDescriptor do
         "HINT_MSG" ->
           "HINT"
 
+        # Attempt messages display the raw student input. For most activities, this is a space-separated
+        # string of selected choice IDs, e.g. "id1 id2 id3." Previously, this code mapped the input
+        # string of ids to the actual content of the selected choices. However, Datashop `<input>`
+        # elements have a maximum character count of 255 chars, so the parsed HTML content of all
+        # selected choices is often too long (e.g. ordering inputs displayed the contents of every
+        # choice). Each activity also has its own "meaning" for student input - e.g. activities are not
+        # required to treat student inputs as lists of choice IDs, so each activity had to be treated
+        # differently to show the human-friendly student "selection." The downside of showing the
+        # raw input string is that the actual choice content seen by the student is not obvious through
+        # Datashop without cross-referencing the IDs.
         "ATTEMPT" ->
-          input = part_attempt.response["input"]
-
-          case activity_type(part_attempt) do
-            # for short answer questions, the input is the text the student entered in the field
-            "oli_short_answer" ->
-              input
-              |> Utils.parse_content()
-
-            # for multiple choice questions, the input is a string id that refers to the selected choice
-            "oli_multiple_choice" ->
-              choices = part_attempt.activity_attempt.transformed_model["choices"]
-              content = Enum.find(choices, &(&1["id"] == input))["content"]
-
-              case content do
-                %{"model" => model} -> Utils.parse_content(model)
-                _ -> Utils.parse_content(content)
-              end
-
-            "oli_check_all_that_apply" ->
-              # CATA Input includes all selected choices as "id1 id2 id3"
-              selected_choices = String.split(input, " ")
-              choices = part_attempt.activity_attempt.transformed_model["choices"]
-
-              contents =
-                choices
-                |> Enum.filter(fn choice ->
-                  Enum.find(
-                    selected_choices,
-                    fn selected_id -> choice["id"] == selected_id end
-                  )
-                end)
-                |> Enum.map(fn choice -> choice["content"] end)
-
-              contents
-              |> Enum.map(fn content_item ->
-                case content_item do
-                  %{"model" => model} -> Utils.parse_content(model)
-                  _ -> Utils.parse_content(content_item)
-                end
-              end)
-
-            # fallback for future activity types
-            _unregistered ->
-              "Input in unregistered activity type: " <> input
-          end
+          part_attempt
+          |> select_input()
+          |> Utils.cdata()
 
         "RESULT" ->
-          content = part_attempt.feedback["content"]
-
-          case content do
-            %{"model" => model} -> Utils.parse_content(model)
-            _ -> Utils.parse_content(content)
-          end
+          part_attempt
+          |> select_feedback()
+          |> Utils.structured_content_to_cdata()
       end
     rescue
       e ->
@@ -107,10 +63,18 @@ defmodule Oli.Analytics.Datashop.Elements.EventDescriptor do
         Type: #{type}, part attempt: #{Kernel.inspect(part_attempt)}"
         The input that created this event could not be found.
         """)
+
+        # Return a generic "Student input" string for datashop if the actual input
+        # cannot be parsed just to give a hint as to what the message is for.
+        "Student input"
     end
   end
 
-  defp activity_type(part_attempt) do
-    part_attempt.activity_attempt.revision.activity_type.slug
+  defp select_input(part_attempt) do
+    part_attempt.response["input"]
+  end
+
+  defp select_feedback(part_attempt) do
+    part_attempt.feedback["content"]["model"]
   end
 end