+++ CHANGELOG.md
@@ -3,9 +3,18 @@
 ## Unreleased
 
 ### Bug Fixes
-- Fix an issue where deleting multiple choice answers could put the question in a state where no incorrect answer is found
+
+- Fix datashop export content model parsing
 
 ### Enhancements
+
+## 0.12.2 (2021-07-21)
+
+### Bug Fixes
+
+- Fix an issue where deleting multiple choice answers could put the question in a state where no incorrect answer is found
+- Fix an issue where activities do not correctly restore their "in-progress" state from student work
+- Fix an issue where images and audio could not be added to activiites
 
 ## 0.12.1 (2021-07-12)